backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic-settings
opencv-python-headless
Pillow
ImageHash
torch
transformers
firebase-admin
google-cloud-aiplatform
google-cloud-firestore
